[move-ide] Refactored parsing symbolicator into parsing analysis
