+++ external-crates/move/crates/move-analyzer/src/analysis/mod.rs
@@ -1,4 +1,5 @@
 // Copyright (c) The Move Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod parsing_analysis;
 pub mod typing_analysis;
